[BSE-1470] Support all numeric arguments for DATEADD/TIMEADD/TIMESTAMPADD (#6627)
* Accept numeric argument for DATE_ADD

* round interval amount in runtime

* Support TIME_ADD

* Fix rounding behavior

* Support TIMESTAMPADD

* Only perform rounding behavior if amount is a float

* Add tests for fractional arguments

* [run CI]

* Add docs for TIMESTAMP_ADD

* [run CI]

* [run CI]

* Use the integer path for none

* [run CI]

+++ BodoSQL/calcite_sql/bodosql-calcite-application/src/main/java/com/bodosql/calcite/application/operatorTables/DatetimeOperatorTable.java
@@ -164,7 +164,7 @@ public final class DatetimeOperatorTable implements SqlOperatorTable {
           OperandTypes.sequence(
                   "DATEADD(UNIT, VALUE, DATETIME)",
                   OperandTypes.ANY,
-                  OperandTypes.INTEGER,
+                  OperandTypes.NUMERIC,
                   OperandTypes.DATETIME)
               .or(OperandTypes.DATETIME_INTERVAL)
               .or(OperandTypes.family(SqlTypeFamily.DATETIME, SqlTypeFamily.INTEGER))
@@ -181,7 +181,7 @@ public final class DatetimeOperatorTable implements SqlOperatorTable {
           OperandTypes.sequence(
               "TIMEADD(UNIT, VALUE, TIME)",
               OperandTypes.ANY,
-              OperandTypes.INTEGER,
+              OperandTypes.NUMERIC,
               OperandTypes.DATETIME),
           SqlFunctionCategory.TIMEDATE);
